requirements.txt
amqp==5.1.1
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
bleach==4.1.0
certifi==2020.12.5
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.4
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.6.0
Deprecated==1.2.13
diff-match-patch==20200713
distlib==0.3.1
Django==3.2
django-3-jet==1.0.7
django-admin-rangefilter==0.6.3
django-allauth==0.45.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-post-office==3.6.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-seed==0.2.2
django-summernote==0.8.11.6
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
eggtestinfo==0.3
et-xmlfile==1.0.1
Faker==8.16.0
filelock==3.0.12
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jsonfield==3.1.0
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.1.0
numpy==1.23.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pendulum==2.1.2
Pillow==8.4.0
prompt-toolkit==3.0.29
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
typing_extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0